backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
scipy==1.11.4
numpy==1.24.3

# AI/ML - Optimized for 4GB VRAM
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
audiocraft==1.3.0
crepe==0.0.13
pretty-midi==0.2.10
music21==9.1.0

# Effects & Mixing (CPU-based where possible)
pedalboard==0.8.0
pyloudnorm==0.1.1

# LLM Integration (GPT-3.5 for speed)
openai==1.3.0
anthropic==0.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Optional: Background processing
# redis==5.0.0
# celery==5.3.0
